% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get_setting}
\alias{get_setting}
\title{Setting getter}
\usage{
get_setting(varstr, value, allow_null = FALSE, param, scope, setter, ...)
}
\arguments{
\item{varstr}{Character. The name of the cached variable.}

\item{value}{Expected value (any type). If provided it overrides the cached value.}

\item{allow_null}{Logical. If \code{FALSE} a helpful error will result if the return
value is \code{NULL}. Otherwise the \code{NULL} will be returned silently.}

\item{param}{Character. Configuration parameter name. See \code{\link[clusterconf]{get_cluster_param}}.}

\item{scope}{Character. Configuration scope. See \code{\link[clusterconf]{get_cluster_param}}}

\item{setter}{Function. The \code{set_*} function used to keep track of the parameter value. This
is needed so that in cases where a parameter is asked for that has not yet been set, the
configuration lookup need only happen once and then the value is cached.}

\item{...}{Passed to \code{\link[clusterconf]{get_cluster_param}}.}
}
\description{
Typcally this should not be used directly. It is relied upon by the \code{get_*}
functions in this package. It is exported mostly to provide transparency around
the available parameters.
}
