% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{hdfs_delete}
\alias{hdfs_delete}
\title{Delete a file or directory}
\usage{
hdfs_delete(path, recursive = FALSE, user = NULL)
}
\arguments{
\item{path}{Character containing file system path}

\item{recursive}{Boolean indicator of whether to recursively delete sub-directories.  Default FALSE.}

\item{user}{Character username to use in WebHDFS operation.  If not provided,
\code{webhdfs.user} will be used and if that has not been set, a call to
\code{\link{guess_user}} will be made.}
}
\value{
Boolean indicator of whether the delete operation succeeded
}
\description{
Shortcut to WebHDFS DELETE operation
}

