% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhdfs_operation.R
\name{hdfs_get}
\alias{hdfs_get}
\title{Generic function to perform HTTP GET request of a WebHDFS operation}
\usage{
hdfs_get(path, operation, return_type = get_return_type())
}
\arguments{
\item{path}{Character containing file system path}

\item{operation}{Character containing WebHDFS operation name}

\item{return_type}{character string. See \code{\link{set_return_type}} for
details and options.}
}
\value{
\code{data.frame} (or other requested type) containing output from
  the WebHDFS operation
}
\description{
Generic function to perform HTTP GET request of a WebHDFS operation
}
\examples{
hdfs_get("/data/raw/tma/SCHEDL", "LISTSTATUS")

}
\seealso{
\url{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html}
  for documentation of WebHDFS REST API commands that can be passed to the
  \code{operation} parameter
}

