% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{hdfs_rename}
\alias{hdfs_rename}
\title{Rename a file or directory}
\usage{
hdfs_rename(old_path, new_path, user = get_user())
}
\arguments{
\item{old_path}{Character path of existing file or directory to be renamed}

\item{new_path}{Character path destination file or directory}

\item{user}{Character username to use in WebHDFS operation.  If not provided,
\code{webhdfs.user} will be used and if that has not been set, a call to
\code{\link{guess_user}} will be made.}
}
\value{
Boolean indicator of whether the rename operation completed successfully
}
\description{
Shortcut to WebHDFS RENAME operation
}
\examples{
\dontrun{
hdfs_rename("/data/old", "/data/new")
}
}
