% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hdfs_timestamp_to_posix}
\alias{hdfs_timestamp_to_posix}
\title{Convert HDFS file timestamps to POSIX}
\usage{
hdfs_timestamp_to_posix(x)
}
\arguments{
\item{x}{Numeric, milliseconds since January 1, 1970 UTC.  Will accept a
character that can be converted to numeric.}
}
\value{
POSIXct formatted timestamp
}
\description{
Convert HDFS file timestamps to POSIX
}
\examples{
hdfs_timestamp_to_posix(1489039200390)

}
