% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namenode_utilities.R
\name{is_namenode_active}
\alias{is_namenode_active}
\title{Test whether the provided namenode is active}
\usage{
is_namenode_active(webhdfs_url)
}
\arguments{
\item{webhdfs_url}{Character containing full WebHDFS URL of the namenode to
test, including the port number and suffix}
}
\value{
Logical indicator of whether the namenode is active
}
\description{
If the supplied namenode is currently a standby, this will detect the
StandbyException from the test operation and return FALSE.  A successful test
operation will cause this function to return TRUE.
}

