% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_cluster.R
\name{set_cluster}
\alias{set_cluster}
\alias{get_cluster}
\title{Specify a Hadoop cluster}
\usage{
set_cluster(name, config_yaml)

get_cluster()
}
\arguments{
\item{name}{Character. Specify the name of the cluster to "point" to.}

\item{config_yaml}{Character. Path to a YAML file specifying cluster settings.}
}
\description{
The easy way to set started - skipping the call to many different \code{set_}
functions.
}
\details{
This function will take either a cluster name or a path to a configuration file.
If a name is provided then a search will be made for a package containing
cluster configurations (named \code{clusterconf.<your cluster name>}). The search
starts in you installed packages (\code{.libPaths()}) and then proceeds to
check your configured repositories (\code{getOption("repos")}).

If you provide a configuration file path instead then that is used in lieu of
searching for a package and the necessary settings are parsed out of it.
}
