% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_cluster.R
\name{set_name_node_url}
\alias{get_name_node_url}
\alias{set_name_node_url}
\title{Manually specify cluster settings for namenode}
\usage{
set_name_node_url(url)

get_name_node_url()
}
\description{
Manually specify cluster settings for namenode
}

