% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_cluster.R
\name{set_webhdfs_port}
\alias{get_webhdfs_port}
\alias{set_webhdfs_port}
\title{Manually specify cluster settings for WebHDFS port number}
\usage{
set_webhdfs_port(port)

get_webhdfs_port()
}
\description{
Manually specify cluster settings for WebHDFS port number
}

