% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_cluster.R
\name{set_webhdfs_suffix}
\alias{get_webhdfs_suffix}
\alias{set_webhdfs_suffix}
\title{Manually specify cluster settings for WebHDFS URL suffix}
\usage{
set_webhdfs_suffix(suffix)

get_webhdfs_suffix()
}
\description{
Manually specify cluster settings for WebHDFS URL suffix
}

