% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unlist_carefully}
\alias{unlist_carefully}
\title{Unlist an object to the innermost \code{data.frame}}
\usage{
unlist_carefully(x)
}
\arguments{
\item{x}{Object returned from call to \code{\link[jsonlite]{fromJSON}},
typically a list}
}
\value{
\code{data.frame} containing the innermost \code{data.frame}-like
  object within \code{x}
}
\description{
Some calls to the WebHDFS REST API return a single JSON object, and some
return an array of JSON objects.  This is a general function that will return
a \code{data.frame} based on the output of either type from
\code{\link[jsonlite]{fromJSON}}
}
\details{
This is similar to simply calling \code{\link{as.data.frame}} on a
\code{list}, but without the nested column names that result.  For example,
where \code{\link{as.data.frame}} could generate a column name of
\code{FileStatuses.FileStatus.accessTime}, this function will result in a
column name of simply \code{accessTime}.
}

