% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{apply_default_configurations}
\alias{apply_default_configurations}
\alias{set_return_type}
\alias{get_return_type}
\alias{set_user}
\alias{get_user}
\alias{guess_user}
\title{WebHDFS Options}
\usage{
apply_default_configurations()

set_return_type(return_type)

get_return_type()

set_user(user = NULL)

get_user()

guess_user()
}
\arguments{
\item{return_type}{character string, must be one of \code{"data.frame", "data.table", "tbl"}}

\item{user}{character string, hadoop user name}
}
\description{
There are many different configurations that are possible. Generally however it is easiest to use
one of the \code{clusterconf.*} packages to manage these. To set any configurations permanently one
can add \code{setOption(param="value")} values in, for example, an .rProfile.
}
\section{options settings}{

In this section the various parameters that are used are defined. Note that all \code{soh}
options will begin with \code{webhdfs.}
\describe{
  \item{webhdfs.return.type}{Specifies the flavor of \code{data.frame} returned when querying
        hive. Most R users are used to working with \code{data.frame} objects. However this function
        allows users to specify a default return type of either \code{data.table} for data.table
        users or \code{tbl} for dplyr users.}
  \item{webhdfs.user}{Some WebHDFS operations and server configurations
        require authentication, and others don't. For those that do it is necessary
        to provide a username as part of the command url. If no username is given
        or an incorrect one is used (i.e., one not associated with an account on
        the hadoop cluster) then some operations (e.g., rename or delete) will fail
        with exit status 1. There is a mechanism by which this package guesses at
        your hadoop username making the assumption that it is the lowercase version
        of the computer user name. This should work for most people, but not all.
        If you are one of the lucky ones maintaining multiple (different) usernames
        then you should set this parameter before trying to use this package.}
}
}

