% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhdfs_operation.R
\name{hdfs_get}
\alias{hdfs_get}
\alias{hdfs_put}
\title{Generic functions to perform HTTP GET or PUT requests of a WebHDFS operation}
\usage{
hdfs_get(path, operation, user = get_user(),
  return_type = get_return_type())

hdfs_put(path, operation, user = get_user())
}
\arguments{
\item{path}{Character containing file system path}

\item{operation}{Character containing WebHDFS operation name}

\item{user}{Character username to use in WebHDFS operation.  If not provided,
\code{webhdfs.user} will be used and if that has not been set, a call to
\code{\link{guess_user}} will be made.}

\item{return_type}{character string. See \code{\link{set_return_type}} for
details and options.}
}
\value{
\code{data.frame} (or other requested type) containing output from
  the WebHDFS operation
}
\description{
The full URL is built from the base WebHDFS URL via the
\code{\link{get_webhdfs_url}} function, and the supplied \code{path} and
\code{operation} parameters.
}
\details{
If the WebHDFS command returns an exception, this function issues a warning
containing the exception message and returns an empty \code{data.frame}.
}
\examples{
\dontrun{
hdfs_get("/data/", "LISTSTATUS")
}

}
\seealso{
\url{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html}
for documentation of WebHDFS REST API commands that can be passed to the
\code{operation} parameter
}
