% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{hdfs_ls}
\alias{hdfs_ls}
\title{List a directory}
\usage{
hdfs_ls(path, recursive = FALSE, return_type = get_return_type())
}
\arguments{
\item{path}{Character containing file system path}

\item{return_type}{character string. See \code{\link{set_return_type}} for
details and options.}

\item{resursive}{Logical indicator of whether to resursively list individual
files within sub-directories.  Default FALSE.}
}
\value{
\code{data.frame} (or other requested type) containing output from
  the WebHDFS operation
}
\description{
Shortcut for WebHDFS LISTSTATUS operation
}
\examples{
hdfs_ls("/data/raw/tma/SCHEDL")

# use wildcard to get new years day from every year
hdfs_ls("/data/raw/tma/SCHEDL/zny/*/01/01")

# use wildcard to get first of the month from every artcc
hdfs_ls("/data/raw/tma/SCHEDL/*/2017/*/01")

}

