% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{hdfs_ls}
\alias{hdfs_ls}
\title{List a directory}
\usage{
hdfs_ls(path, recursive = FALSE, return_type = get_return_type())
}
\arguments{
\item{path}{Character containing file system path}

\item{return_type}{character string. See \code{\link{set_return_type}} for
details and options.}

\item{resursive}{Logical indicator of whether to resursively list individual
files within sub-directories.  Default FALSE.}
}
\value{
\code{data.frame} (or other requested type) containing output from
  the WebHDFS operation
}
\description{
Shortcut for WebHDFS LISTSTATUS operation
}
\details{
Wildcards (*) in the provided \code{path} are handled by getting all contents at that
level of the directory structure, and appending any remaining portions of the
\code{path} below that level.
}
\examples{
\dontrun{
# basic usage to list directory contents
hdfs_ls("/data/")

# recursively list files within sub-directories
hdfs_ls("/data/blah/2016", recursive = TRUE)

# use wildcard to get the first day of every month in 2016
hdfs_ls("/data/blah/2016/*/01")

# use wildcards to get all days in July in every year
hdfs_ls("/data/blah/*/07/*")
}


}

