% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.R
\name{hdfs_ls}
\alias{hdfs_ls}
\title{Generic function to perform HTTP GET request of a WebHDFS operation}
\usage{
hdfs_ls(path, return_type = get_return_type())
}
\arguments{
\item{path}{Character containing file system path}

\item{return_type}{character string. See \code{\link{set_return_type}} for
details and options.}
}
\value{
\code{data.frame} (or other requested type) containing output from
  the WebHDFS operation
}
\description{
Generic function to perform HTTP GET request of a WebHDFS operation
}
\examples{
hdfs_ls("/data/raw/tma/SCHEDL")

}

