% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_directory.R
\name{hdfs_makedir}
\alias{hdfs_makedir}
\title{Make a directory}
\usage{
hdfs_makedir(path, user = NULL, permission = NULL)
}
\arguments{
\item{path}{Character path to create}

\item{user}{Character username to use in WebHDFS operation.  If not provided,
\code{webhdfs.user} will be used and if that has not been set, a call to
\code{\link{guess_user}} will be made.}

\item{permission}{Optional octal specification of permissions for the new
directory.  If none is provided, the WebHDFS default (755) is used.}
}
\value{
Boolean indicator of whether the directory was created successfully
}
\description{
Shortcut to WebHDFS MKDIRS operation
}

